requirements.txt
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.8.0
altair==4
tiktoken==0.6.0
dependency-injector==4.41.0
# uncomment to use huggingface llms
huggingface-hub==0.14.1
python-multipart==0.0.5
confluent-kafka==2.3.0
# uncomment to use instructor embeddings
# InstructorEmbedding==1.0.1
# sentence-transformers==2.2.2
fastapi~=0.109.0
uvicorn==0.18.3
pydantic~=1.10.14
aiokafka~=0.10.0
motor~=3.4.0
pymongo~=4.6.3
langchain~=0.1.16